Sequential code update working

+++ formula.h
@@ -2,8 +2,8 @@
 #define FORMULA_H_
 #include "grayscale.h"
 
-double convolution(double* image, int mask[3][3], int row, int col, struct dimensions d);
+int convolution(int* image, int mask[3][3], int row, int col, struct dimensions d);
 
-double* sobel_edge_detection(double* grayscaleImg, struct dimensions dim);
+int* sobel_edge_detection(int* grayscaleImg, struct dimensions dim);
 
 #endif
