+++ grayscale.h
@@ -9,8 +9,8 @@ struct dimensions
 
 struct dimensions bmp_dimensions(char *file_name);
 
-double * bmp_to_grayscale(char *file_name);
+int * bmp_to_grayscale(char *file_name);
 
-void matrix_to_file(char *file_name, double *matrix);
+void matrix_to_file(char *file_name, int *matrix);
 
 #endif
